Make backend enhancements
- Check if directory folder exists prior to deletion
- Add slug to return object for when listing created
- Add gallery id to both listing and user objects upon creation

+++ core/roommates/models/galleries.py
@@ -85,5 +85,7 @@ class Gallery(models.Model):
             self.uuid
         )
 
-        full_path = os.path.abspath(dir_path)    
-        shutil.rmtree(full_path)
+        full_path = os.path.abspath(dir_path)
+        
+        if os.path.isdir(full_path):    
+            shutil.rmtree(full_path)
